+++ 8_bit_synth/8_bit_synth/calculate.h
@@ -23,17 +23,17 @@ enum usu
     PUSH_SUCCESS
 };
 
-int calculate(int num1,int num2,char c);
-int cal(char * exp);
-int priority(char c);
-void initialize_stack(int *stack,int *top);
-void initialize_stack2(char *stack,int *top);
-int is_stack_full(int *top);
-int is_stack_empty(int *top);
-int push_stack(int *stack,int *num,int *top);
-int push_stack2(char *stack,char *c,int *top);
-int pop_stack(int *stack,int *top);
-char pop_stack2(char *stack,int *top);
-int compare_priority(char *str,char *stack2,int *top2,int *stack1,int *top1);
+long int calculate(long int num1,long int num2,char c);
+long int cal(char * exp);
+long int priority(char c);
+void initialize_stack(long int *stack,long int *top);
+void initialize_stack2(char *stack,long int *top);
+long int is_stack_full(long int *top);
+long int is_stack_empty(long int *top);
+long int push_stack(long int *stack,long int *num,long int *top);
+long int push_stack2(char *stack,char *c,long int *top);
+long int pop_stack(long int *stack,long int *top);
+char pop_stack2(char *stack,long int *top);
+long int compare_priority(char *str,char *stack2,long int *top2,long int *stack1,long int *top1);
 
 #endif /* calculate_h */
